+++ server/api/settings/[id].delete.ts
@@ -1,10 +1,11 @@
-import { PrismaClient } from '@prisma/client';
-import { getValidatedIdParam } from '../utils';
+import { PrismaClient } from "@prisma/client";
+import { getCaller, getValidatedIdParam } from "../utils";
 
 const prisma = new PrismaClient();
 
 export default defineEventHandler(async (event) => {
-  const id = getValidatedIdParam(event, 'setting');
-  await prisma.setting.delete({ where: { id } });
+  const id = getValidatedIdParam(event, "setting");
+  const { id: callerId } = getCaller(event);
+  await prisma.setting.delete({ where: { id, userId: callerId } });
   return { success: true };
-}); 
+});
